changing database structure and add payment gatewat

+++ bookings/views.py
@@ -19,10 +19,10 @@ from mainapp.models import Login
 from django.contrib import messages
 # useremail
 
-
+# this is the booking form page....
 @never_cache
 def bookings(request, id):
-    if f"user_{id}_uname" not in request.session or f"user_{id}_upass" not in request.session or f"user_{id}_uemail" not in request.session:
+    if f"user_{id}_uname" not in request.session and f"user_{id}_upass" not in request.session and f"user_{id}_uemail" not in request.session:
         return HttpResponseRedirect('/login/')
     elif 'user_{}_uname'.format(id) in request.session and 'user_{}_upass'.format(id) in request.session and 'user_{}_uemail'.format(id) in request.session:
         data1 = {}
@@ -35,11 +35,13 @@ def bookings(request, id):
             hstate1 = request.GET.get('hstate')
             hcost1 = request.GET.get('hcost')
             image = request.GET.get('image_url')
+            
             # url = f"/dashboard/{id}"
             url = "/dashboard/{}".format(id)
             data1 = {'un1': un1, 'pw1': password1, 'em1': email1, 'hname1': hname1,
                      'hcity1': hcity1, 'hstate1': hstate1, 'hcost1': hcost1, 'url': url, 'id': id, 'image': image}
             return render(request, 'booking.html', data1)
+        
         try:
             if request.method == "POST":
                 name = request.POST.get('name')
@@ -72,17 +74,13 @@ def bookings(request, id):
                 image = request.GET.get('image_url')
 
                 if start < str(date.today()):
-                    messages.warning(
-                        request, 'Your starting date must be equal or more than today !')
+                    messages.warning(request, 'Your starting date must be equal or more than today !')
                     url = f"/dashboard/{id}"
-
                     data1 = {'un1': un1, 'pw1': password1, 'em1': email1, 'hname1': hname1,
                              'hcity1': hcity1, 'hstate1': hstate1, 'hcost1': hcost1, 'url': url, 'id': id, 'image': image}
                     return render(request, 'booking.html', data1)
                 elif start > end:
-                    messages.warning(
-                        request, 'your ending date must be more than start date !')
-
+                    messages.warning(request, 'your ending date must be more than start date !')
                     url = "/dashboard/{}".format(id)
                     data1 = {'un1': un1, 'pw1': password1, 'em1': email1, 'hname1': hname1,
                              'hcity1': hcity1, 'hstate1': hstate1, 'hcost1': hcost1, 'url': url, 'id': id, 'image': image}
@@ -93,18 +91,20 @@ def bookings(request, id):
                                         userpassword=password, start=start, end=end, hotelname=hotelname, city=hotelcity, state=hotelstate, current_cost=hotelcost)
                     data.save()
                     url = "/dashboard/{}".format(id)
-                    full_name = (name+last).upper()
+                    full_name = (name+' '+ last).upper()
                     data1 = {'url': url, 'id': id}
+                    
                     # this is for redirecting into the dashboard page ...
                     response = HttpResponseRedirect(url)
-                    messages.success(
-                        request, f'Your booking has been done for mr. {full_name}. You can see your order details below !')
+                    messages.success(request, f'Your booking has been done for mr. {full_name}. You can see your order details below !')
                     return response
         except Exception as e:
             pass
     return render(request, 'booking.html', data1)
 
 
+
+# this is dashboard page.....
 @never_cache
 def dashboard(request, id):
     if 'user_{}_uname'.format(id) not in request.session and 'user_{}_upass'.format(id) not in request.session and 'user_{}_uemail'.format(id) not in request.session:
@@ -115,8 +115,7 @@ def dashboard(request, id):
         useremail = request.session['user_{}_uemail'.format(id)]
         password = request.session['user_{}_upass'.format(id)]
         if Login.objects.filter(username=username, email=useremail, password=password).exists():
-            tabel = Bookinghotel.objects.filter(
-                username=username, userpassword=password)
+            tabel = Bookinghotel.objects.filter( username=username, userpassword=password)
             hotelurl = '/hotellist/{}/{}'.format('all', id)
             reviewurl = '/review/{}'.format(id)
             # this url is for dashboard page itself ....
@@ -127,12 +126,11 @@ def dashboard(request, id):
         return render(request, 'dashboard.html', data)
 
 
-# this is for order details page..
-
+# this is for order details page.....
 @never_cache
 def details(request):
     id = request.GET.get('session__id')
-    if 'user_{}_uname'.format(id) not in request.session and 'user_{}_upass'.format(id) not in request.session and 'user_{}_uemail'.format(id) not in request.session and 'user_{}_uemail'.format(id) not in request.session:
+    if 'user_{}_uname'.format(id) not in request.session and 'user_{}_upass'.format(id) not in request.session and 'user_{}_uemail'.format(id) not in request.session:
         return HttpResponseRedirect('/login/')
     elif 'user_{}_uname'.format(id) in request.session and 'user_{}_upass'.format(id) in request.session and 'user_{}_uemail'.format(id) in request.session:
         data = {}
@@ -148,8 +146,7 @@ def details(request):
             bool = (maindata.payment_status.lower() == 'unpaid')
             start = str(maindata.start)
             end = str(maindata.end)
-            res = (dt.strptime(end, "%Y-%m-%d") -
-                   dt.strptime(start, "%Y-%m-%d")).days
+            res = (dt.strptime(end, "%Y-%m-%d") - dt.strptime(start, "%Y-%m-%d")).days
             total_cost = res*maindata.current_cost*maindata.no_rooms
             data = {'un': un, 'pw': password, 'uemail': uemail,
                     'maindata': maindata, 'url': url, 'cost': total_cost, 'id': id, 'bool': bool, 'order_id': detail_id, 'bool': bool}
@@ -158,14 +155,13 @@ def details(request):
 
 # in each page context we have to pass the id value because it is the session key variable so we use this so that we can check if someone is logged in or logged out ...
 
-# this page is for deleting the order...
-
-# for deleting the order ...
 
 
+
+# this page is for deleting the order....
 def delete(request):
     id = request.GET.get('session__id')
-    if 'user_{}_uname'.format(id) in request.session and 'user_{}_upass'.format(id) in request.session:
+    if 'user_{}_uname'.format(id) in request.session and 'user_{}_upass'.format(id) in request.session and 'user_{}_uemail'.format(id) in request.session:
         id1 = request.GET.get('id1')
         id = request.GET.get('session__id')
         Bookinghotel.objects.filter(id=id1).delete()
